refactor: Update hooks
