+++ packages/ui-components/src/index.tsx
@@ -2,5 +2,6 @@ import "./styles.css";
 
 export * from "./Templates";
 export * from "./components";
+export * from "./hooks/useScreen";
 
 export * as utils from "./utils";
