+++ packages/web-app/lib/constants.tsx
@@ -1,5 +1,5 @@
 export const NATION = "0x333A4823466879eeF910A04D473505da62142069";
-export const frameworkAddress = "0x51b024Ca13F6E044df4932431bF8DD0E5d4b81ba";
-export const arbitratorAddress = "0xA723Fc96d9180637B21048168D0344CC677da64c";
+export const frameworkAddress = "0xb2f61CA5eF1cb524b86bc7994F54C9f6A069dCf1";
+export const arbitratorAddress = "0xb2f61CA5eF1cb524b86bc7994F54C9f6A069dCf1";
 export const cohortAddress = "0xD86B638A810fc0C811c6A27152744eEFFE7262C0";
 export const safeTxServiceUrl = "https://safe-transaction-goerli.safe.global";
